pyproject.toml
[project]
name = "predictive-maintenance-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "autogen-agentchat>=0.7.5",
    "autogen-ext[openai]>=0.7.5",
    "crewai>=1.6.1",
    "crewai-tools[mcp]>=1.6.1",
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "langchain>=1.2.10",
    "langchain-openai>=1.1.11",
    "langfuse>=4.0.0",
    "langgraph>=1.0.10",
    "matplotlib>=3.10.8",
    "mcp>=1.26.0",
    "numpy>=2.4.3",
    "openai>=2.26.0",
    "openai-agents>=0.12.3",
    "pandas>=3.0.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
]
